+++ src/main/java/org/stellar/sdk/SorobanServer.java
@@ -368,7 +368,40 @@ public class SorobanServer implements Closeable {
    */
   public Transaction prepareTransaction(Transaction transaction)
       throws IOException, SorobanRpcErrorResponse, PrepareTransactionException {
-    SimulateTransactionResponse simulateTransactionResponse = this.simulateTransaction(transaction);
+    SimulateTransactionResponse simulateTransactionResponse = simulateTransaction(transaction);
+    return prepareTransaction(transaction, simulateTransactionResponse);
+  }
+
+  /**
+   * Prepare the transaction using the simulateTransaction obtained by the user in advance, apply
+   * the simulateTransaction results to a new copy of the transaction which is then returned.
+   * Setting the ledger footprint and authorization, so the resulting transaction is ready for
+   * signing and sending.
+   *
+   * <p>The returned transaction will also have an updated fee that is the sum of fee set on
+   * incoming transaction with the contract resource fees estimated from simulation. It is advisable
+   * to check the fee on returned transaction and validate or take appropriate measures for
+   * interaction with user to confirm it is acceptable.
+   *
+   * @param transaction The transaction to prepare. It should include exactly one operation, which
+   *     must be one of {@link InvokeHostFunctionOperation}, {@link
+   *     BumpFootprintExpirationOperation}, or {@link RestoreFootprintOperation}. Any provided
+   *     footprint will be ignored. You can use {@link Transaction#isSorobanTransaction()} to check
+   *     if a transaction is a Soroban transaction. Any provided footprint will be overwritten.
+   *     However, if your operation has existing auth entries, they will be preferred over ALL auth
+   *     entries from the simulation. In other words, if you include auth entries, you don't care
+   *     about the auth returned from the simulation. Other fields (footprint, etc.) will be filled
+   *     as normal.
+   * @param simulateTransactionResponse The {@link SimulateTransactionResponse} to use for preparing
+   *     the transaction.
+   * @return Returns a copy of the {@link Transaction}, with the expected authorizations (in the
+   *     case of invocation) and ledger footprint added. The transaction fee will also automatically
+   *     be padded with the contract's minimum resource fees discovered from the simulation.
+   * @throws PrepareTransactionException If preparing the transaction fails.
+   */
+  public Transaction prepareTransaction(
+      Transaction transaction, SimulateTransactionResponse simulateTransactionResponse)
+      throws PrepareTransactionException {
     if (simulateTransactionResponse.getError() != null) {
       throw new PrepareTransactionException(
           "simulation transaction failed, the response contains error information.",